Support library 28, AGP 3.3.1, Kotlin 1.3.21, Gradle 5.2.1
Removed deprecated code unused in support lib 28. Fixed compile errors.
Fixed code in IconifiedTabLayout to work with design lib 28.

+++ commons/src/main/java/net/xpece/android/net/ConnectivityInfo.kt
@@ -1,6 +1,6 @@
 package net.xpece.android.net
 
-class ConnectivityInfo(@ConnectivityReceiver.State val state: Long, val isAirplaneModeEnabled: Boolean) {
+class ConnectivityInfo(@ConnectivityReceiver.State val state: Int, val isAirplaneModeEnabled: Boolean) {
 
     companion object {
         @JvmStatic private val CONNECTED = ConnectivityInfo(ConnectivityReceiver.Companion.STATE_CONNECTED, false)
@@ -21,7 +21,7 @@ class ConnectivityInfo(@ConnectivityReceiver.State val state: Long, val isAirpla
     val isDisconnected: Boolean
         get() = state == ConnectivityReceiver.STATE_DISCONNECTED
 
-    fun copy(@ConnectivityReceiver.State state: Long = this.state, isAirplaneModeEnabled: Boolean = this.isAirplaneModeEnabled): ConnectivityInfo {
+    fun copy(@ConnectivityReceiver.State state: Int = this.state, isAirplaneModeEnabled: Boolean = this.isAirplaneModeEnabled): ConnectivityInfo {
         if (!isAirplaneModeEnabled) {
             when (state) {
                 ConnectivityReceiver.STATE_CONNECTED -> return CONNECTED
